pms-backend/requirements.txt
# ── Framework & Server ───────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# ── Database ─────────────────────────────────────────────
sqlalchemy==2.0.35
alembic==1.13.3
aiosqlite==0.20.0

# ── Auth ─────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# ── Validation ───────────────────────────────────────────
pydantic==2.9.0
pydantic-settings==2.5.0
email-validator==2.2.0

# ── Date utils ───────────────────────────────────────────
python-dateutil==2.9.0

# ── Environment ──────────────────────────────────────────
python-dotenv==1.0.1
